+++ app/src/main/java/com/app/namllahprovider/data/api/order/change_order_status/ChangeOrderRequest.kt
@@ -7,7 +7,7 @@ import okhttp3.RequestBody
 data class ChangeOrderRequest(
     val orderId: Int,
     val orderStatusRequestType: OrderStatusRequestType,
-    val estimatedTime: Int = 0,
+    val estimatedTime: Double = 0.0,
     val amount: Int = 0,
     val cancelReasonId: Int = 0,
     val estimatedPriceParts: Double = 0.0,
